requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==2.11.7
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
dill==0.3.6
Django==3.2
django-bower==5.2.0
django-ckeditor==6.2.0
django-extensions==3.2.3
django-js-asset==2.0.0
django-manifest-loader==1.0.0
django-mathfilters==1.0.0
django-taggit==4.0.0
djangorestframework==3.11.2
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
isort==5.11.5
lazy-object-proxy==1.9.0
lxml==4.9.2
mccabe==0.7.0
numpy==1.21.6
oscrypto==1.3.0
pandas==1.1.5
Pillow==9.5.0
pip==22.3.1
platformdirs==3.1.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pylint==2.13.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pypdf==3.9.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.6
webencodings==0.5.1
wheel==0.38.4
wincertstore==0.2
wrapt==1.15.0
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
